c: 1 cpp: 48, 57, 75, 342, 419, 455, 2410

+++ 1.cpp
@@ -11,7 +11,7 @@ public:
     for (auto i = 0; i < ranges::ssize(nums); i++) {
       auto num = nums[i];
       if (indexForNum.contains(target - num)) {
-        return {indexForNum[target - num], static_cast<int>(i)};
+        return {indexForNum[target - num], i};
       }
       indexForNum[num] = i;
     }
